requirements.txt
constructs>=10.0.0,<11.0.0
boto3
pytest
black
pylint
aws-cdk-lib==2.141.0
